+++ src/main/java/com/daml/extensions/damlmavenplugin/CodeGen.java
@@ -4,17 +4,14 @@
  */
 package com.daml.extensions.damlmavenplugin;
 
-import java.io.IOException;
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLClassLoader;
-import java.nio.file.Path;
 import java.util.List;
 import java.util.Optional;
-import java.util.Scanner;
 
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.repository.ArtifactRepository;
@@ -131,20 +128,10 @@ public class CodeGen extends MojoBase {
     }
 
     private String getDamlVersion() throws MojoFailureException {
-        String errorMsg =
-                "Cannot determine project sdk version. Make sure that `daml.yaml` includes a line specifying `sdk-version`.";
-
-        Path damlYaml = getDamlYamlFile();
-        try (Scanner scanner = new Scanner(damlYaml)) {
-            while (scanner.hasNextLine()) {
-                String line = scanner.nextLine().trim();
-                if (line.startsWith("sdk-version:")) {
-                    return line.substring(12).trim();
-                }
-            }
-        } catch (IOException e) {
-            throw new MojoFailureException(errorMsg, e);
+        String sdkVersion = DamlProject.create().getSdkVersion();
+        if (sdkVersion == null || sdkVersion.isEmpty()) {
+            throw new MojoFailureException("Cannot determine project sdk version. Make sure that `daml.yaml` includes a line specifying `sdk-version`.");
         }
-        throw new MojoFailureException(errorMsg);
+        return sdkVersion;
     }
 }