+++ src/main/java/com/daml/extensions/damlmavenplugin/DamlProject.java
@@ -8,6 +8,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -15,18 +16,35 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import org.apache.maven.plugin.MojoFailureException;
 import org.yaml.snakeyaml.Yaml;
 
 public class DamlProject {
 
+    private static String DAML_YAML_NAME = "daml.yaml";
+
     private final Path damlYaml;
     private final Map<String, Object> project;
 
-    public static DamlProject create(Path damlYaml) throws IOException {
+    public static DamlProject create(Path damlYaml) throws MojoFailureException {
         Yaml yaml = new Yaml();
         try (InputStream is = Files.newInputStream(damlYaml)) {
             return new DamlProject(damlYaml, yaml.load(is));
+        } catch (IOException ioe) {
+            throw new MojoFailureException("Error reading daml.yaml", ioe);
         }
+    }
+
+    public static DamlProject create() throws MojoFailureException {
+        Path cwd = Paths.get("").toAbsolutePath();
+        while (cwd != null) {
+            Path damlYaml = cwd.resolve(DAML_YAML_NAME);
+            if (Files.exists(damlYaml)) {
+                return create(damlYaml);
+            }
+            cwd = cwd.getParent();
+        }
+        throw new MojoFailureException("daml.yaml file not found");
     }
 
     private DamlProject(Path damlYaml, Map<String, Object> project) {
@@ -36,6 +54,10 @@ public class DamlProject {
 
     private Path projectRelative(String p) {
         return damlYaml.getParent().resolve(p);
+    }
+
+    public String getSdkVersion() {
+        return (String) project.get("sdk-version");
     }
 
     @SuppressWarnings("unchecked")